Account

+++ database/migrations/updates/2024_02_06_173021_drop_some_cols_from_users_table.php
@@ -13,10 +13,25 @@ return new class extends Migration
     {
         Schema::table('users', function (Blueprint $table) {
 
-            $table->dropColumn('custom_field_1');
-            $table->dropColumn('custom_field_2');
-            $table->dropColumn('social_media_1');
-            $table->dropColumn('social_media_2');
+            if (Schema::hasColumn('users', 'custom_field_1')) {
+
+                $table->dropColumn('custom_field_1');
+            }
+
+            if (Schema::hasColumn('users', 'custom_field_2')) {
+
+                $table->dropColumn('custom_field_2');
+            }
+
+            if (Schema::hasColumn('users', 'social_media_1')) {
+                
+                $table->dropColumn('social_media_1');
+            }
+
+            if (Schema::hasColumn('users', 'social_media_2')) {
+
+                $table->dropColumn('social_media_2');
+            }
         });
     }
